Adding file issue service
